==> Project 1/1. Data_Collection_Processing/P1_twitter-target-term_FINAL.ipynb <==
import pandas as pd 
import requests as req
import json 
import datetime as dt
import numpy as np
import seaborn as sns
import matplotlib as plt

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
analyzer = SentimentIntensityAnalyzer()

json_NYSE = ('NYSE.json')
df_NYSE = pd.read_json(json_NYSE)
#df_NYSE.tail()

text_sentiment_compound_NYSE  = []
for i in df_NYSE['text']:
    compound = analyzer.polarity_scores(i)['compound']
    text_sentiment_compound_NYSE.append(compound)

df_NYSE['Compound_NYSE']=text_sentiment_compound_NYSE
#df_NYSE.info()

df_NYSE['Date']=df_NYSE['timestamp'].map(lambda x: x.date())

#use the rename(columns={})
df_NYSE_cleaned_headers = df_NYSE[['Date','Compound_NYSE']]

#get rid of all compounds with a score of equal to zero
df_NYSE_cleaned_headers = df_NYSE_cleaned_headers.loc[(df_NYSE_cleaned_headers['Compound_NYSE']>0) | (df_NYSE_cleaned_headers['Compound_NYSE'] <0)]

df_NYSE_cleaned_compound_NYSE = df_NYSE_cleaned_headers.reset_index(drop=True)

mean_compound_NYSE=df_NYSE_cleaned_compound_NYSE.groupby("Date")["Compound_NYSE"].mean()

#mean_compound_NYSE
df_mean_compound_NYSE=pd.DataFrame(mean_compound_NYSE)
df_mean_compound_NYSE.to_csv("NYSE.csv")
#df_mean_compound_NYSE

json_NASDAQ = ('NASDAQ.json')
df_NASDAQ = pd.read_json(json_NASDAQ)
#df_NASDAQ.tail()

text_sentiment_compound_NASDAQ  = []
for i in df_NASDAQ['text']:
    compound = analyzer.polarity_scores(i)['compound']
    text_sentiment_compound_NASDAQ.append(compound)

df_NASDAQ['Compound_NASDAQ']=text_sentiment_compound_NASDAQ
#df_NASDAQ.info()

df_NASDAQ['Date']=df_NASDAQ['timestamp'].map(lambda x: x.date())

#use the rename(columns={})
df_NASDAQ_cleaned_headers = df_NASDAQ[['Date','Compound_NASDAQ']]

#get rid of all compounds with a score of equal to zero
df_NASDAQ_cleaned_headers = df_NASDAQ_cleaned_headers.loc[(df_NASDAQ_cleaned_headers['Compound_NASDAQ']>0) | (df_NASDAQ_cleaned_headers['Compound_NASDAQ'] <0)]

df_NASDAQ_cleaned_compound_NASDAQ = df_NASDAQ_cleaned_headers.reset_index(drop=True)

mean_compound_NASDAQ=df_NASDAQ_cleaned_compound_NASDAQ.groupby("Date")["Compound_NASDAQ"].mean()

#mean_compound_NASDAQ
df_mean_compound_NASDAQ=pd.DataFrame(mean_compound_NASDAQ)
df_mean_compound_NASDAQ.to_csv("NASDAQ.csv")
#df_mean_compound_NASDAQ

json_DowJones = ('DowJones.json')
df_DowJones = pd.read_json(json_DowJones)
#df_NASDAQ.tail()

text_sentiment_compound_DowJones  = []
for i in df_DowJones['text']:
    compound = analyzer.polarity_scores(i)['compound']
    text_sentiment_compound_DowJones.append(compound)

df_DowJones['Compound_DowJones']=text_sentiment_compound_DowJones
#df_DowJones.info()

df_DowJones['Date']=df_DowJones['timestamp'].map(lambda x: x.date())

#use the rename(columns={})
df_DowJones_cleaned_headers = df_DowJones[['Date','Compound_DowJones']]

#get rid of all compounds with a score of equal to zero
df_DowJones_cleaned_headers = df_DowJones_cleaned_headers.loc[(df_DowJones_cleaned_headers['Compound_DowJones']>0) | (df_DowJones_cleaned_headers['Compound_DowJones'] <0)]

df_DowJones_cleaned_compound_DowJones = df_DowJones_cleaned_headers.reset_index(drop=True)

mean_compound_DowJones=df_DowJones_cleaned_compound_DowJones.groupby("Date")["Compound_DowJones"].mean()

#mean_compound_DowJones
df_mean_compound_DowJones=pd.DataFrame(mean_compound_DowJones)
df_mean_compound_DowJones.to_csv("DowJones.csv")
#df_mean_compound_DowJones

json_stockmarket = ('stockmarket.json')
df_stockmarket = pd.read_json(json_stockmarket)
#df_stockmarket.tail()

text_sentiment_compound_stockmarket  = []
for i in df_stockmarket['text']:
    compound = analyzer.polarity_scores(i)['compound']
    text_sentiment_compound_stockmarket.append(compound)

df_stockmarket['Compound_stockmarket']=text_sentiment_compound_stockmarket
#df_stockmarket.info()

df_stockmarket['Date']=df_stockmarket['timestamp'].map(lambda x: x.date())

#use the rename(columns={})
df_stockmarket_cleaned_headers = df_stockmarket[['Date','Compound_stockmarket']]

#get rid of all compounds with a score of equal to zero
df_stockmarket_cleaned_headers = df_stockmarket_cleaned_headers.loc[(df_stockmarket_cleaned_headers['Compound_stockmarket']>0) | (df_stockmarket_cleaned_headers['Compound_stockmarket'] <0)]

df_stockmarket_cleaned_compound_stockmarket = df_stockmarket_cleaned_headers.reset_index(drop=True)

mean_compound_stockmarket=df_stockmarket_cleaned_compound_stockmarket.groupby("Date")["Compound_stockmarket"].mean()

#mean_compound_stockmarket
df_mean_compound_stockmarket=pd.DataFrame(mean_compound_stockmarket)
df_mean_compound_stockmarket.to_csv("stockmarket.csv")
#df_mean_compound_stockmarket

# Compining CSV files. Getting all mean compound for target terms on one table 
file_stockmarket = "stockmarket.csv"
file_DowJones = "DowJones.csv"
file_NASDAQ = "NASDAQ.csv"
file_NYSE = "NYSE.csv"

stockmarket = pd.read_csv(file_stockmarket)
DowJones = pd.read_csv(file_DowJones)
NASDAQ = pd.read_csv(file_NASDAQ)
NYSE = pd.read_csv(file_NYSE)

stockmarket_DowJones = pd.merge(stockmarket, DowJones, how="inner", on="Date")
stockmarket_DowJones_NASDAQ = pd.merge(stockmarket_DowJones,NASDAQ , how="inner", on="Date")
stockmarket_DowJones_NASDAQ_NYSE = pd.merge(stockmarket_DowJones_NASDAQ, NYSE, how="inner", on="Date")
terms_stocks1 = stockmarket_DowJones_NASDAQ_NYSE.drop(stockmarket_DowJones_NASDAQ_NYSE.index[[1,2,8,9]])
terms_stocks1 = terms_stocks1.rename(columns={"Compound_DowJones":"target dow","Compound_NASDAQ":"target nasdaq","Compound_NYSE":"target nyse"}).reset_index(drop=True)
#stocks1 = stocks1.drop(stocks1.index[10]).reset_index(drop=True)

terms_stocks = terms_stocks1.drop(terms_stocks1.index[[0,10]]).reset_index(drop=True)

terms_stocks

terms_stocks1.to_csv("target_terms_sameday.csv")

terms_stocks2 = terms_stocks1.drop(terms_stocks1.index[[9,10]]).rename(columns = {"target dow": "target dow prior","target nasdaq": "target nasdaq prior","target nyse": "target nyse prior"})
terms_stocks2 = terms_stocks2.drop(['Date','Compound_stockmarket'], axis=1)
terms_stocks2

final_stock_terms = pd.concat([terms_stocks, terms_stocks2], axis=1, join_axes=[terms_stocks.index])
final_stock_terms = final_stock_terms.drop(['Compound_stockmarket'], axis=1)
final_stock_terms

final_stock_terms.to_csv("target_terms.csv")
