==> Project 1/1. Data_Collection_Processing/P2_Twitter_scrapper_FINAL.ipynb <==
import pandas as pd 
import requests as req
import json 
import datetime as datetime
import numpy as np
import seaborn as sns
import matplotlib as plt

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
analyzer = SentimentIntensityAnalyzer()

json = ('NovDecDataUser.json')

df = pd.read_json(json)

df.tail()

df['fullname'].unique()

text_sentiment_compound  = []
for i in df['text']:
    compound = analyzer.polarity_scores(i)['compound']
    text_sentiment_compound.append(compound)

df['Compound']=text_sentiment_compound

df.info()

df['Date']=df['timestamp'].map(lambda x: x.date())

df['UserName'] = df['fullname']

#use the rename(columns={})
df_cleaned_headers = df[['UserName','Date','Compound']]

#get rid of all compounds with a score of equal to zero
df_cleaned_headers = df_cleaned_headers.loc[(df_cleaned_headers['Compound']>0) | (df_cleaned_headers['Compound'] <0)]

df_cleaned_compound = df_cleaned_headers.reset_index(drop=True)

from pandas_datareader import data
import matplotlib.pyplot as plt
import warnings

df_cleaned_compound['Date'] = pd.to_datetime(df_cleaned_compound['Date'])
df_cleaned_compound.head()

df_cleaned_compound = df_cleaned_compound.set_index("Date")

user_dated = df_cleaned_compound.ix['2017-12-01':'2017-12-14']

user_dated.head()

user_dated.groupby(["UserName"])["Compound"].mean()
user_dated.to_csv("TargetUsers.csv")


user_dated.info()
