==> Project 1/1. Data_Collection_Processing/P3_Finance2_FINAL.ipynb <==
** FIRST STEP IS TO INSTALL DEPENDENCIES **

#Dependencies
#! pip install pandas-datareader
from pandas_datareader import data
import matplotlib.pyplot as plt
import pandas as pd
import warnings

** Next step provides selected closing index values.  This step uses datareader.  **

#Choose the ticker symbols for analysis

tickers = ['^IXIC', '^DJI', '^GSPC', '^RUT', '^GDAXI', '^XAX']

# Define which online source one should use
data_source = 'yahoo'

# We would like all available data from 01/01/2000 until 12/31/2016.
start_date = '2017-11-14'
end_date = '2017-12-15'

# User pandas_reader.data.DataReader to load the desired data. As simple as that.
panel_data = data.DataReader(tickers, data_source, start_date, end_date)

# Getting just the adjusted closing prices. This will return a Pandas DataFrame
# The index in this DataFrame is the major index of the panel_data.
adj_close = panel_data.ix['Adj Close']

# Getting all weekdays between 01/01/2000 and 12/31/2016
all_weekdays = pd.date_range(start=start_date, end=end_date, freq='B')

# How do we align the existing prices in adj_close with our new set of dates?
# All we need to do is reindex adj_close using all_weekdays as the new index
adj_close = adj_close.reindex(all_weekdays)

# Reindexing will insert missing values (NaN) for the dates that were not present
# in the original set. To cope with this, we can fill the missing by replacing them
# with the latest available price for each instrument.
adj_close = adj_close.fillna(method='ffill')

** The index for the adjusted closing table is currently the date.  Need to change this for later steps. **

adj_close = adj_close.reset_index()
adj_close
#adj_close

** To put performance of the various indices on the same scale - calculate one day returns (day (t) / day (t-1) - 1) - No value for the first day with index values - set it to 0 - will eliminate that row when you combine with the twitter data.  Should make sure the first day of finance data is earlier than first date of twitter data. **


#Calculate one day returns

count = 0
daily_DJI = []
daily_GSPC = []
daily_IXIC = []
daily_RUT = []
#daily_BTC = []
daily_GDAXI = []
daily_XAX = []

for index, row in adj_close.iterrows():
   
        if count == 0:
            Dailydji = 0
            Dailygspc = 0
            Dailyixic = 0
            Dailyrut = 0
            #Dailybtc = 0
            Dailygdaxi = 0
            Dailyxax = 0
            count = count + 1
            daily_DJI.append(Dailydji)
            daily_GSPC.append(Dailygspc)
            daily_IXIC.append(Dailyixic)
            daily_RUT.append(Dailyrut)
            daily_GDAXI.append(Dailygdaxi)
            daily_XAX.append(Dailyxax)
            #daily_BTC.append(Dailybtc)
            #print (count)
        else:
        
            Dailydji = adj_close["^DJI"][count]/adj_close["^DJI"][count - 1] - 1
            Dailygspc = adj_close["^GSPC"][count]/adj_close["^GSPC"][count - 1] - 1
            Dailyixic = adj_close["^IXIC"][count]/adj_close["^IXIC"][count - 1] - 1
            Dailyrut = adj_close["^RUT"][count]/adj_close["^RUT"][count - 1] - 1
            Dailygdaxi = adj_close["^GDAXI"][count]/adj_close["^GDAXI"][count - 1] - 1
            Dailyxax = adj_close["^XAX"][count]/adj_close["^XAX"][count - 1] - 1
            #Dailybtc = adj_close["BTC-USD"][count]/adj_close["BTC-USD"][count-1] -1
            count = count + 1
            daily_DJI.append(Dailydji)
            daily_GSPC.append(Dailygspc)
            daily_IXIC.append(Dailyixic)
            daily_RUT.append(Dailyrut)
            #daily_BTC.append(Dailybtc)
            daily_GDAXI.append(Dailygdaxi)
            daily_XAX.append(Dailyxax)
            #adj_close.loc[index,'Daily'] = adj_close["^DJI"][count]/adj_close["^DJI"][count - 1] - 1
            #print (count)
            
        #print("index is" + str(index))
        #print("row is" + str(row))

daily_DJI = pd.DataFrame({"Daily DJI":daily_DJI})
daily_GSPC = pd.DataFrame({"Daily GSPC":daily_GSPC})
daily_IXIC = pd.DataFrame({"Daily IXIC":daily_IXIC})
daily_RUT = pd.DataFrame({"Daily RUT":daily_RUT})
daily_GDAXI = pd.DataFrame({"Daily GDAXI":daily_GDAXI})
daily_XAX = pd.DataFrame({"Daily XAX":daily_XAX})
#daily_BTC = pd.DataFrame({"Daily BTC":daily_BTC})
merged = adj_close.join(daily_RUT,how = "inner")
merged = merged.join(daily_DJI,how = "inner")
merged = merged.join(daily_GSPC,how = "inner")
merged = merged.join(daily_IXIC,how = "inner")
merged = merged.join(daily_GDAXI,how = "inner")
merged = merged.join(daily_XAX,how = "inner")
merged.rename(columns={"index":"Date"},inplace=True)

merged = merged.sort_values(by='Date').drop(merged.index[0:13])

merged = merged.reset_index(drop=True)
merged

** Next - import the target user dataset that Ethan provided.  Group by date and user.  Want to line up twitter sentiment and index performance by date.  Need to make sure they are lined up if you are using a generic index.  Use the default date format for datetime in order to standardize for merging tables later.**

target_users = pd.read_csv("TargetUsers.csv")

target_users = target_users.groupby(["Date","UserName"]).mean()
target_users = target_users.reset_index()

target_users_CNN_Money = target_users.loc[target_users['UserName'] == 'CNNMoney']
target_users_WSJ_Markets = target_users.loc[target_users['UserName'] == 'WSJ Markets']
target_users_CNBC = target_users.loc[target_users['UserName'] == 'CNBC']
target_users_Financial_Times = target_users.loc[target_users['UserName'] == 'Financial Times']
target_users_Bloomberg_Markets = target_users.loc[target_users['UserName'] == 'Bloomberg Markets']
target_users_Justin_Wolfers = target_users.loc[target_users['UserName'] == 'Justin Wolfers']
target_users_Paul_Krugman = target_users.loc[target_users['UserName'] == 'Paul Krugman']
target_users_Jim_Cramer = target_users.loc[target_users['UserName'] == 'Jim Cramer']
target_users_Motley_Fool = target_users.loc[target_users['UserName'] == 'The Motley Fool']

from datetime import datetime
target_users_CNN_Money['Date'] = pd.to_datetime(target_users_CNN_Money['Date'])
target_users_WSJ_Markets['Date'] = pd.to_datetime(target_users_WSJ_Markets['Date'])
target_users_CNBC['Date'] = pd.to_datetime(target_users_CNBC['Date'])
target_users_Financial_Times['Date'] = pd.to_datetime(target_users_Financial_Times['Date'])
target_users_Bloomberg_Markets['Date'] = pd.to_datetime(target_users_Bloomberg_Markets['Date'])
target_users_Justin_Wolfers['Date'] = pd.to_datetime(target_users_Justin_Wolfers['Date'])
target_users_Paul_Krugman['Date'] = pd.to_datetime(target_users_Paul_Krugman['Date'])
target_users_Jim_Cramer['Date'] = pd.to_datetime(target_users_Jim_Cramer['Date'])
target_users_Motley_Fool['Date'] = pd.to_datetime(target_users_Motley_Fool['Date'])


target_users_CNN_Money

** Now combine the twitter data with the market data into one final table called combined_table.  **

combined_table = pd.merge(target_users_CNN_Money, merged, on="Date", how="inner")
combined_table = combined_table.drop(['UserName','^DJI',"^GDAXI","^GSPC","^IXIC","^RUT","^XAX"], axis=1)
combined_table.rename(columns={"Compound":"CNN Sentiment"},inplace=True)

combined_table = pd.merge(target_users_WSJ_Markets, combined_table, on="Date", how="inner")
combined_table = combined_table.drop(['UserName'], axis=1)
combined_table.rename(columns={"Compound":"WSJ Sentiment"},inplace=True)

combined_table = pd.merge(target_users_CNBC, combined_table, on="Date", how="inner")
combined_table = combined_table.drop(['UserName'], axis=1)
combined_table.rename(columns={"Compound":"CNBC Sentiment"},inplace=True)

combined_table = pd.merge(target_users_Financial_Times, combined_table, on="Date", how="inner")
combined_table = combined_table.drop(['UserName'], axis=1)
combined_table.rename(columns={"Compound":"Fin Times Sentiment"},inplace=True)

combined_table = pd.merge(target_users_Bloomberg_Markets, combined_table, on="Date", how="inner")
combined_table = combined_table.drop(['UserName'], axis=1)
combined_table.rename(columns={"Compound":"Bloomberg Sentiment"},inplace=True)

combined_table = pd.merge(target_users_Justin_Wolfers, combined_table, on="Date", how="outer")
combined_table = combined_table.drop(['UserName'], axis=1)
combined_table.rename(columns={"Compound":"Justin Wolfers Sentiment"},inplace=True)

combined_table = pd.merge(target_users_Paul_Krugman, combined_table, on="Date", how="inner")
combined_table = combined_table.drop(['UserName'], axis=1)
combined_table.rename(columns={"Compound":"Paul Krugman Sentiment"},inplace=True)

combined_table = pd.merge(target_users_Jim_Cramer, combined_table, on="Date", how="inner")
combined_table = combined_table.drop(['UserName'], axis=1)
combined_table.rename(columns={"Compound":"Jim Cramer Sentiment"},inplace=True)

combined_table = pd.merge(target_users_Motley_Fool, combined_table, on="Date", how="inner")
combined_table = combined_table.drop(['UserName'], axis=1)
combined_table.rename(columns={"Compound":"Motley Fool Sentiment"},inplace=True)

combined_table = combined_table.sort_values(by='Date')
combined_table = combined_table.reset_index(drop=True)
combined_table = combined_table.drop(combined_table.index[[1,2,8,9]])
combined_table = combined_table.reset_index(drop=True)
combined_table = combined_table.fillna(0)

combined_table2 = combined_table.iloc[0:10,:]
combined_table2 = combined_table2.drop(combined_table2.index[0]).reset_index(drop=True)
combined_table2.info()


combined_table1 = combined_table.iloc[0:9,0:10]
combined_table1 = combined_table1.reset_index(drop=True)
combined_table1 = combined_table1.rename(columns={"Motley Fool Sentiment":"prior motley","Jim Cramer Sentiment":"prior jim","Paul Krugman Sentiment":"prior paul","Justin Wolfers Sentiment":"prior justin","Bloomberg Sentiment":"prior bloomberg","Fin Times Sentiment":"prior ft","CNBC Sentiment":"prior cnbc","WSJ Sentiment":"prior wsj","CNN Sentiment":"prior cnn"})
combined_table1 = combined_table1.drop(['Date'], axis=1)
combined_table1.info()

final_table = pd.concat([combined_table2, combined_table1], axis=1, join_axes=[combined_table2.index])
final_table.head()

final_table.info()

** Now for some factor generation.  Start by lining up market performance with prior day's tweet sentiment.  Want to see if you can use tweet sentiment to predict the market.  Use this for an investment strategy.  **

table = pd.read_csv("target_terms.csv")
#final_table.to_csv('combined_table2.csv')
#combined_table = pd.read_csv("combined_table2.csv")
table = table.drop(['Unnamed: 0','Date'], axis=1)
table

final_table = pd.concat([final_table, table], axis=1, join_axes=[final_table.index])
final_table.to_csv('combined_table.csv')
final_table.info()

final_table

** Now calculate the correlation matrix - and export it to "correlation.csv". **

corr = final_table.corr()
corr.to_csv('correlation_table.csv')
corr
